v2.63.0
Official public ShaderOptimizer API

So a few commits ago I changed the namespace of the Optimizer and I've come to realize a lot of tools used functions of that.
I'm sorry for breaking stuff.

I've added public functions marked with [PublicAPI] that I am committed to keep working as long as possible. This means no random namespace renames, method signature changes etc.

If your tool requires any other API functions please DM me or make a github issue, so we can add official public support for it.
More importantly, if one of my future changes breaks stuff, please, please let me know as fast as possible. In most cases I can and will reverse it.
In this case reversing would cause a lot of confusion as some tools already adapted their approach.
