+++ Editor/UnlockedMaterialList.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 79f6717198499b9449876e0244d39822
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
